+++ frontend/src/components/Login/TownSelectionPart2.test.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable no-await-in-loop,@typescript-eslint/no-loop-func,no-restricted-syntax */
 import React from 'react'
+import '@testing-library/jest-dom'
 import { ChakraProvider } from '@chakra-ui/react'
 import { fireEvent, render, RenderResult, waitFor, within } from '@testing-library/react'
 import userEvent, { TargetElement } from '@testing-library/user-event'
@@ -103,194 +104,10 @@ function wrappedTownSelection() {
     <TownSelection doLogin={doLoginMock}/></CoveyAppContext.Provider></ChakraProvider>;
 }
 
-describe('Part 1 - Public room listing', () => {
-  beforeEach(() => {
-    jest.useFakeTimers();
-    mocklistTowns.mockReset();
-  })
-  it('is called when rendering (hopefully by a useeffect, this will be checked manually)', async () => {
-    mocklistTowns.mockImplementation(() => listTowns(nanoid()));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(1);
-    })
-    renderData.unmount();
-  });
-  it('updates every 2000 msec', async () => {
-    mocklistTowns.mockImplementation(() => listTowns(nanoid()));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(1);
-    })
-    jest.advanceTimersByTime(2000);
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(2);
-    })
-    jest.advanceTimersByTime(1000);
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(2);
-    })
-    renderData.unmount();
-  });
-  it('stops updating when unmounted', async () => {
-    mocklistTowns.mockImplementation(() => listTowns(nanoid()));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(1);
-    })
-    jest.advanceTimersByTime(2000);
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(2);
-    })
-    renderData.unmount();
-    jest.advanceTimersByTime(10000);
-    await waitFor(() => {
-      expect(mocklistTowns)
-        .toBeCalledTimes(2);
-    })
-  });
-  it('updates the page with all rooms stored in currentPublicTowns', async () => {
-    const suffix1 = nanoid();
-    const suffix2 = nanoid();
-    const expectedTowns1 = await listTowns(suffix1)
-    const expectedTowns2 = await listTowns(suffix2)
-    mocklistTowns.mockImplementation(() => listTowns(suffix1));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expectedTowns1.rooms.map((town) => expect(renderData.getByText(town.friendlyName))
-        .toBeInTheDocument());
-    })
-    mocklistTowns.mockImplementation(() => listTowns(suffix2));
-    jest.advanceTimersByTime(2000);
-    await waitFor(() => {
-      expectedTowns2.rooms.forEach((town) => expect(renderData.getByText(town.friendlyName))
-        .toBeInTheDocument());
-      expectedTowns1.rooms.forEach((town) => expect(renderData.queryByText(town.friendlyName))
-        .not
-        .toBeInTheDocument());
-    })
-
-  });
-  it('does not include the hardcoded demo in the listing', async () => {
-    const suffix = nanoid();
-    const expectedTowns1 = await listTowns(suffix)
-    mocklistTowns.mockImplementation(() => listTowns(suffix));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expectedTowns1.rooms.map((town) => expect(renderData.getByText(town.friendlyName))
-        .toBeInTheDocument());
-    })
-    expect(renderData.queryByText('demoTownName'))
-      .not
-      .toBeInTheDocument();
-  });
-  it('sorts rooms by occupancy descending', async () => {
-    const suffix1 = nanoid();
-    const suffix2 = nanoid();
-    const expectedTowns1 = await listTowns(suffix1)
-    expectedTowns1.rooms = expectedTowns1.rooms.sort((a, b) => b.currentOccupancy - a.currentOccupancy);
-
-    const expectedTowns2 = await listTowns(suffix2)
-    expectedTowns2.rooms = expectedTowns2.rooms.sort((a, b) => b.currentOccupancy - a.currentOccupancy);
-
-    mocklistTowns.mockImplementation(() => listTowns(suffix1));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expectedTowns1.rooms.map((town) => expect(renderData.getByText(town.friendlyName))
-        .toBeInTheDocument());
-    })
-    // All rooms are in doc, now make sure they are sorted by occupancy
-    let rows = renderData.getAllByRole('row');
-    for (let i = 1; i < rows.length; i += 1) { // off-by-one for the header row
-      // console.log(rows[i]);
-      const existing = within(rows[i])
-        .getByText(expectedTowns1.rooms[i - 1].friendlyName);
-      expect(existing)
-        .toBeInTheDocument();
-      for (let j = 0; j < expectedTowns1.rooms.length; j += 1) {
-        if (j !== i - 1) {
-          expect(within(rows[i])
-            .queryByText(expectedTowns1.rooms[j].friendlyName))
-            .not
-            .toBeInTheDocument();
-        }
-      }
-    }
-    // Now, do that all again to make sure it sorts EVERY run
-    mocklistTowns.mockImplementation(() => listTowns(suffix2));
-    jest.advanceTimersByTime(2000);
-    await waitFor(() =>
-      expectedTowns2.rooms.map((town) => expect(renderData.getByText(town.friendlyName))
-        .toBeInTheDocument())
-    )
-
-    // All rooms are in doc, now make sure they are sorted by occupancy
-    rows = renderData.getAllByRole('row');
-    for (let i = 1; i < rows.length; i += 1) { // off-by-one for the header row
-      // console.log(rows[i]);
-      const existing = within(rows[i])
-        .getByText(expectedTowns2.rooms[i - 1].friendlyName);
-      expect(existing)
-        .toBeInTheDocument();
-      for (let j = 0; j < expectedTowns2.rooms.length; j += 1) {
-        if (j !== i - 1) {
-          expect(within(rows[i])
-            .queryByText(expectedTowns2.rooms[j].friendlyName))
-            .not
-            .toBeInTheDocument();
-        }
-      }
-    }
-  });
-  it('represents each row in the table as specified', async () => {
-    const suffix1 = nanoid();
-    const expectedTowns = await listTowns(suffix1);
-    expectedTowns.rooms = expectedTowns.rooms.sort((a, b) => b.currentOccupancy - a.currentOccupancy);
-    mocklistTowns.mockImplementation(() => listTowns(suffix1));
-    const renderData = render(wrappedTownSelection());
-    await waitFor(() => {
-      expectedTowns.rooms.forEach((town) => expect(renderData.getByText(town.friendlyName))
-        .toBeInTheDocument());
-    })
-    const rows = renderData.getAllByRole('row');
-    expectedTowns.rooms.forEach((town) => {
-      const row = rows.find(each => within(each)
-        .queryByText(town.coveyRoomID));
-      if (row) {
-        const cells = within(row)
-          .queryAllByRole('cell');
-        // Cell order: friendlyName, roomID, occupancy/join + button
-        expect(cells.length)
-          .toBe(3);
-        expect(within(cells[0])
-          .queryByText(town.friendlyName))
-          .toBeInTheDocument();
-        expect(within(cells[1])
-          .queryByText(town.coveyRoomID))
-          .toBeInTheDocument();
-        expect(within(cells[2])
-          .queryByText(`${town.currentOccupancy}/${town.maximumOccupancy}`))
-          .toBeInTheDocument();
-      } else {
-        fail(`Could not find row for town ${town.coveyRoomID}`);
-      }
-    })
-  });
-});
-
 describe('Town Selection - depends on Part 1 passing', () => {
   let renderData: RenderResult<typeof import("@testing-library/dom/types/queries")>;
   let townIDToJoinField: HTMLInputElement;
-  let newTownNameField: HTMLInputElement;
-  let newTownIsPublicCheckbox: HTMLInputElement;
   let userNameField: HTMLInputElement;
-  let newTownButton: TargetElement;
   let joinTownByIDButton: TargetElement;
   let expectedTowns: TownListResponse;
 
@@ -309,11 +126,8 @@ describe('Town Selection - depends on Part 1 passing', () => {
     renderData = render(wrappedTownSelection());
     await waitFor(() => expect(renderData.getByText(`room1${suffix}`))
       .toBeInTheDocument());
-    newTownIsPublicCheckbox = renderData.getByLabelText('Publicly Listed') as HTMLInputElement;
     townIDToJoinField = renderData.getByPlaceholderText('ID of town to join, or select from list') as HTMLInputElement;
-    newTownNameField = renderData.getByPlaceholderText('New Town Name') as HTMLInputElement;
     userNameField = renderData.getByPlaceholderText('Your name') as HTMLInputElement;
-    newTownButton = renderData.getByTestId('newTownButton');
     joinTownByIDButton = renderData.getByTestId('joinTownByIDButton');
   });
   describe('Part 2 - Joining existing towns', () => {
@@ -661,182 +475,5 @@ describe('Town Selection - depends on Part 1 passing', () => {
       });
     });
   });
-
-  describe('Part 3 - Creating a new town', () => {
-
-    const createTownWithOptions = async (params: { townName: string, userName: string, togglePublicBox?: boolean, townID?: string, roomPassword?: string, errorMessage?: string }) => {
-      fireEvent.change(userNameField, { target: { value: params.userName } });
-      await waitFor(() => {
-        expect(userNameField.value)
-          .toBe(params.userName);
-      });
-      fireEvent.change(newTownNameField, { target: { value: params.townName } });
-      await waitFor(() => expect(newTownNameField.value)
-        .toBe(params.townName));
-      if (params.togglePublicBox) {
-        fireEvent.click(newTownIsPublicCheckbox);
-        await waitFor(() => expect(newTownIsPublicCheckbox.checked)
-          .toBe(false));
-      }
-      mockCreateTown.mockReset();
-      if (params.townID && params.roomPassword) {
-        mockCreateTown.mockReturnValue({
-          coveyRoomID: params.townID,
-          coveyRoomPassword: params.roomPassword
-        });
-      } else if (params.errorMessage) {
-        mockCreateTown.mockRejectedValue(new Error(params.errorMessage));
-      } else {
-        fail('Invalid config');
-      }
-      userEvent.click(newTownButton);
-    }
-    describe('when clicking create', () => {
-      describe('with invalid values', () => {
-        it('displays an error toast "Unable to create town" if the username is empty', async () => {
-          await createTownWithOptions({
-            userName: '',
-            townName: nanoid(),
-            errorMessage: 'FAIL'
-          })
-          await waitFor(() => expect(mockToast)
-            .toBeCalledWith({
-              title: 'Unable to create town',
-              description: 'Please select a username before creating a town',
-              status: 'error'
-            }));
-        });
-        it('displays an error toast "Unable to create town" if the newTownName is empty', async () => {
-          await createTownWithOptions({
-            townName: '',
-            userName: nanoid(),
-            errorMessage: 'FAIL'
-          })
-          await waitFor(() => expect(mockToast)
-            .toBeCalledWith({
-              title: 'Unable to create town',
-              description: 'Please enter a town name',
-              status: 'error'
-            }));
-        });
-      });
-      describe('with valid values', () => {
-
-        it('calls createTown on the apiClient with the provided values (public town)', async () => {
-          const townID = nanoid();
-          const roomPassword = nanoid();
-          const townName = nanoid();
-          await createTownWithOptions({
-            townName,
-            userName: nanoid(),
-            townID,
-            roomPassword
-          })
-          await waitFor(() => expect(mockCreateTown)
-            .toBeCalledWith({
-              friendlyName: townName,
-              isPubliclyListed: true
-            }));
-        });
-
-        it('calls createTown on the apiClient with the provided values (not public town)', async () => {
-          const townID = nanoid();
-          const roomPassword = nanoid();
-          const townName = nanoid();
-          await createTownWithOptions({
-            townName,
-            userName: nanoid(),
-            townID,
-            roomPassword,
-            togglePublicBox: true
-          })
-          await waitFor(() => expect(mockCreateTown)
-            .toBeCalledWith({
-              friendlyName: townName,
-              isPubliclyListed: false
-            }));
-
-        });
-
-        it('displays a toast "Town newTownName is ready to go!" when successful', async () => {
-          const townID = nanoid();
-          const roomPassword = nanoid();
-          const townName = nanoid();
-          await createTownWithOptions({
-            townName,
-            userName: nanoid(),
-            townID,
-            roomPassword,
-            togglePublicBox: true
-          })
-          await waitFor(() => expect(mockCreateTown)
-            .toBeCalledWith({
-              friendlyName: townName,
-              isPubliclyListed: false
-            }));
-          await waitFor(() => expect(mockToast)
-            .toBeCalledWith(expect.objectContaining({
-              title: `Town ${townName} is ready to go!`,
-              status: 'success',
-              isClosable: true,
-              duration: null
-            })));
-        });
-        it('after success, calls Video.setup, doLogin, and connect with the entered username and newly generated coveyRoomID', async () => {
-          const townID = nanoid();
-          const roomPassword = nanoid();
-          const userName = nanoid();
-          const townName = nanoid();
-
-          // Configure mocks
-          mockVideoSetup.mockReset();
-          const videoToken = nanoid();
-          mockVideoSetup.mockReturnValue(Promise.resolve({ providerVideoToken: videoToken }))
-          doLoginMock.mockReset();
-          doLoginMock.mockReturnValue(Promise.resolve(true));
-
-          // Create town
-          await createTownWithOptions({
-            townName,
-            userName,
-            townID,
-            roomPassword,
-            togglePublicBox: true
-          })
-
-          // Check for call sequence
-          await waitFor(() => expect(mockVideoSetup)
-            .toBeCalledWith(userName, townID));
-          await waitFor(() => expect(doLoginMock)
-            .toBeCalledWith({ providerVideoToken: videoToken }));
-          await waitFor(() => expect(mockConnect)
-            .toBeCalledWith(videoToken));
-
-        });
-        it('displays an error toast "Unable to connect to Towns Service" if an error occurs in createTown', async () => {
-          const errorMessage = `Oops#${nanoid()}`;
-          const townName = nanoid();
-          await createTownWithOptions({
-            townName,
-            userName: nanoid(),
-            errorMessage
-          })
-          await waitFor(() => expect(mockCreateTown)
-            .toBeCalledWith({
-              friendlyName: townName,
-              isPubliclyListed: true
-            }));
-          await waitFor(() => expect(mockToast)
-            .toBeCalledWith({
-              title: 'Unable to connect to Towns Service',
-              status: 'error',
-              description: `Error: ${errorMessage}`
-            }));
-        });
-
-      });
-    });
-  });
-
 });
