+++ frontend/src/components/Login/TownSettings.test.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { ChakraProvider } from '@chakra-ui/react';
 import React from 'react';
+import '@testing-library/jest-dom'
 import { fireEvent, render, RenderResult, waitFor } from '@testing-library/react';
 import { nanoid } from 'nanoid';
 import { TargetElement } from '@testing-library/user-event';
